refactor logFileStore
